+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.6.0 - 2017-01-05
+
+- BBox reduced decimal precision to 6
+- Lat & Lng decimal toFixed(6)
+- Meters toFixed(1)
+
 ## 1.5.0 - 2016-12-25
 
 - Add bboxToCenter & hash function